Split CSV parsing into nicer interface
