+++ include/updater/csv_source.hpp
@@ -1,0 +1,19 @@
+#ifndef OSRM_UPDATER_CSV_SOURCE_HPP
+#define OSRM_UPDATER_CSV_SOURCE_HPP
+
+#include "updater/source.hpp"
+
+namespace osrm
+{
+namespace updater
+{
+namespace csv
+{
+    SegmentLookupTable readSegmentValues(const std::vector<std::string> &paths);
+    TurnLookupTable readTurnValues(const std::vector<std::string> &paths);
+}
+}
+}
+
+
+#endif
